Add helpful string on canvas, for WebGL2 only demos (#744)
* Add helpful string on canvas, for WebGL2 only demos

+++ examples/core/transform/app.js
@@ -1,7 +1,7 @@
 /* eslint-disable camelcase */
 /* global window */
-import {AnimationLoop, Buffer, Model, pickModels, picking} from 'luma.gl';
-import {Transform} from 'luma.gl';
+import {AnimationLoop, Buffer, Model, pickModels, picking, Transform, isWebGL2} from 'luma.gl';
+import {Log} from 'probe.gl';
 
 const RED = new Uint8Array([255, 0, 0, 255]);
 
@@ -16,6 +16,9 @@ const INFO_HTML = `
   </a>
 `;
 /* eslint-enable max-len */
+
+// Text to be displayed on environments when this demos is not supported.
+const ALT_TEXT = 'THIS DEMO REQUIRES WEBLG2, BUT YOUR BRWOSER DOESN\'T SUPPORT IT';
 
 const EMIT_VS = `\
 #version 300 es
@@ -122,8 +125,10 @@ void main()
 `;
 
 const NUM_INSTANCES = 1000;
+const log = new Log({id: 'transform'}).enable();
 
 let pickPosition = [0, 0];
+let isDemoSupported = true;
 function mousemove(e) {
   pickPosition = [e.offsetX, e.offsetY];
 }
@@ -134,7 +139,7 @@ function mouseleave(e) {
 const animationLoop = new AnimationLoop({
   glOptions: {
     webgl2: true,
-    webgl1: false,
+    webgl1: true,
     debug: true
   },
 
@@ -142,6 +147,11 @@ const animationLoop = new AnimationLoop({
 
   /* eslint-disable max-statements */
   onInitialize({canvas, gl}) {
+    isDemoSupported = isWebGL2(gl);
+    if (!isDemoSupported) {
+      log.error(ALT_TEXT)();
+      return {isDemoSupported};
+    }
     gl.canvas.addEventListener('mousemove', mousemove);
     gl.canvas.addEventListener('mouseleave', mouseleave);
 
@@ -219,7 +229,8 @@ const animationLoop = new AnimationLoop({
       colorBuffer,
       offsetBuffer,
       renderModel,
-      transform
+      transform,
+      isDemoSupported
     };
   },
   /* eslint-enable max-statements */
@@ -237,6 +248,9 @@ const animationLoop = new AnimationLoop({
     time
   }) {
 
+    if (!isDemoSupported) {
+      return;
+    }
     transform.run({
       uniforms: {
         u_time: time
@@ -283,12 +297,22 @@ const animationLoop = new AnimationLoop({
   },
 
   onFinalize({renderModel, transform}) {
-    renderModel.delete();
-    transform.delete();
+    if (renderModel) {
+      renderModel.delete();
+    }
+    if (transform) {
+      transform.delete();
+    }
   }
 });
 
 animationLoop.getInfo = () => INFO_HTML;
+animationLoop.isSupported = () => {
+  return isDemoSupported;
+}
+animationLoop.getAltText = () => {
+  return ALT_TEXT;
+}
 
 export default animationLoop;
